[upload] - fix issue with uploading indicatorFields
