+++ demisto_sdk/commands/content_graph/objects/indicator_field.py
@@ -1,17 +1,14 @@
-from typing import Callable, Optional, Set
-
-import demisto_client
-from pydantic import Field
+from typing import Optional, Set
 
 from demisto_sdk.commands.common.constants import MarketplaceVersions
 from demisto_sdk.commands.content_graph.common import ContentType
-from demisto_sdk.commands.content_graph.objects.content_item import ContentItem
+from demisto_sdk.commands.content_graph.objects.indicator_incident_field import (
+    IndicatorIncidentField,
+)
 
 
-class IndicatorField(ContentItem, content_type=ContentType.INDICATOR_FIELD):  # type: ignore[call-arg]
-    cli_name: str = Field(alias="cliName")
+class IndicatorField(IndicatorIncidentField, content_type=ContentType.INDICATOR_FIELD):  # type: ignore[call-arg]
     type: str
-    associated_to_all: bool = Field(alias="associatedToAll")
 
     def summary(
         self,
@@ -23,16 +20,4 @@ class IndicatorField(ContentItem, content_type=ContentType.INDICATOR_FIELD):  # 
         return summary
 
     def metadata_fields(self) -> Set[str]:
-        return (
-            super()
-            .metadata_fields()
-            .union(
-                {
-                    "type",
-                }
-            )
-        )
-
-    @classmethod
-    def _client_upload_method(cls, client: demisto_client) -> Callable:
-        return client.import_incident_fields
+        return super().metadata_fields().union({"type"})
